Add RobotsState and movement logic/collision detection
Added empty and start pieces to enum and represented
Board and RobotsState as two different "layers" of the state.
The Board keeps track of the immutable information
(blocking pieces, goal and start). The RobotsState keeps track
of the movable pieces.

Update Board toString to also output the RobotState if provided.